refactor subnav for logged in routes
