+++ app/frontend/i18n/i18n.ts
@@ -113,6 +113,10 @@ i18n.use(initReactI18next).init({
           title: "Digital Permit Tool",
           description: "Lorem ipsum here is the site description",
           keywords: "BC, british columba, permit, portal, hub, permitting, permit application",
+          activePermits: "Active Permits",
+          approvedPermits: "Approved Permits",
+          myAccount: "My Account",
+          giveFeedback: "Give Feedback",
         },
       },
     },
